Rename object CCScheduler
