+++ modules/effects/arrow-effects-rx2/src/main/kotlin/arrow/effects/ObservableK.kt
@@ -8,7 +8,7 @@ import arrow.typeclasses.Applicative
 import io.reactivex.Observable
 import io.reactivex.ObservableEmitter
 import kotlin.coroutines.experimental.CoroutineContext
-import arrow.effects.ContextScheduler.asScheduler
+import arrow.effects.CoroutineContextScheduler.asScheduler
 
 fun <A> Observable<A>.k(): ObservableK<A> = ObservableK(this)
